+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @drupal-js-sdk/core
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @drupal-js-sdk/xhr@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes