+++ packages/drupal-js-sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # drupal-js-sdk
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @drupal-js-sdk/xhr@0.5.2
+  - @drupal-js-sdk/auth@0.5.2
+  - @drupal-js-sdk/core@0.5.2
+  - @drupal-js-sdk/menu@0.5.2
+  - @drupal-js-sdk/role@0.5.2
+  - @drupal-js-sdk/settings@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes