+++ packages/menu/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @drupal-js-sdk/menu
 
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @drupal-js-sdk/xhr@0.5.2
+  - @drupal-js-sdk/core@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes